Add option to use xcodebuild instead of xctool.
Adds the ability to pass `-x` or `--xcodebuild` to `objc` to use
xcodebuild instead of xctool. In order to maintain rough parity with
xctool's prettified output, adds xcpretty as a dependency.

The option is helpful for users of up-to-date versions of Xcode because
xctool's compatibility with Xcode lags behind a few versions. Xcode is
currently at 8.3.1; xctool's compatibility extends only to 8.1.